Flatten character classes into one single GDKCharacter
Outstanding bug: Mesh visibility no longer switches back to TP mesh on death.
